fix: widen glyph offsets to UINT32
The proportional 32px font bitmap grew past 64 KiB, so glyph offsets
no longer fit UINT16 and clang rejected the constant conversion.
Make SFB_FONT_GLYPH.Offset (and the lookup outputs) UINT32.

+++ submodules/uefi/edk2/QcomModulePkg/Application/LinuxLoader/SuperFbFont.c
@@ -9,7 +9,7 @@
 
 BOOLEAN
 SfbFontGetGlyph (IN CHAR16 Ch,
-                 OUT UINT16 *Offset,
+                 OUT UINT32 *Offset,
                  OUT UINT8  *Width,
                  OUT UINT8  *Advance)
 {
